What if we did this instead?
It looks like code for listening to our own `Source.schema` doesn't
belong in `_addSourceListeners`. I think this code does the same thing,
but is also kind of silly because it goes through ChildProcess
unnecessarily. I'm not totally clear why we are listening to the
Source.schema from a different process anyway; shouldn't the
Source.schema from the `fetching` process be synchronized to mongo in
the same way that the one from the `api` process is?