Stop two tests from accidentally saving reports
We have to remove all the reports in these two for a very sneaky
reason. If you run either of these test files alone, there's no
problem; no reports get created.  But when you use mocha to run this
file and one which imports '../lib/fetching/report-writer', the bots
will start adding reports to the queue and they'll get saved. With
enough latency, the fix here actually won't be enough to stop these
reports from interacting with other tests, as this test file could
exit with reports still in the queue. The moral is probably that
sharing state with `require` statements isn't very good for modular
testing.